feat(vald-operator): consolidate ValdRelease component methods and tests into single package using oapi-codegen generated types
Collapse the per-component subpackages (agent, defaults, discoverer, gateway,
manager) into the single valdrelease package, moving their setter methods and
tests alongside the oapi-codegen generated flat types, and regenerate the
values schema accordingly.
